requirements.txt
Django==1.9
djangosaml2==0.17.2
